refactor: move shell and snapshot code out of codex-core
