+++ codex-rs/core/src/tools/handlers/shell_tests.rs
@@ -9,7 +9,6 @@ use crate::exec_env::create_env;
 use crate::sandboxing::SandboxPermissions;
 use crate::shell::Shell;
 use crate::shell::ShellType;
-use crate::shell_snapshot::ShellSnapshot;
 use crate::tools::context::FunctionToolOutput;
 use crate::tools::context::ToolInvocation;
 use crate::tools::context::ToolPayload;
@@ -22,42 +21,25 @@ use codex_shell_command::powershell::try_find_powershell_executable_blocking;
 use codex_shell_command::powershell::try_find_pwsh_executable_blocking;
 use serde_json::json;
 use tokio::sync::Mutex;
-use tokio::sync::watch;
 
 /// The logic for is_known_safe_command() has heuristics for known shells,
 /// so we must ensure the commands generated by [ShellCommandHandler] can be
 /// recognized as safe if the `command` is safe.
 #[test]
 fn commands_generated_by_shell_command_handler_can_be_matched_by_is_known_safe_command() {
-    let bash_shell = Shell {
-        shell_type: ShellType::Bash,
-        shell_path: PathBuf::from("/bin/bash"),
-        shell_snapshot: crate::shell::empty_shell_snapshot_receiver(),
-    };
+    let bash_shell = Shell::new(ShellType::Bash, PathBuf::from("/bin/bash"));
     assert_safe(&bash_shell, "ls -la");
 
-    let zsh_shell = Shell {
-        shell_type: ShellType::Zsh,
-        shell_path: PathBuf::from("/bin/zsh"),
-        shell_snapshot: crate::shell::empty_shell_snapshot_receiver(),
-    };
+    let zsh_shell = Shell::new(ShellType::Zsh, PathBuf::from("/bin/zsh"));
     assert_safe(&zsh_shell, "ls -la");
 
     if let Some(path) = try_find_powershell_executable_blocking() {
-        let powershell = Shell {
-            shell_type: ShellType::PowerShell,
-            shell_path: path.to_path_buf(),
-            shell_snapshot: crate::shell::empty_shell_snapshot_receiver(),
-        };
+        let powershell = Shell::new(ShellType::PowerShell, path.to_path_buf());
         assert_safe(&powershell, "ls -Name");
     }
 
     if let Some(path) = try_find_pwsh_executable_blocking() {
-        let pwsh = Shell {
-            shell_type: ShellType::PowerShell,
-            shell_path: path.to_path_buf(),
-            shell_snapshot: crate::shell::empty_shell_snapshot_receiver(),
-        };
+        let pwsh = Shell::new(ShellType::PowerShell, path.to_path_buf());
         assert_safe(&pwsh, "ls -Name");
     }
 }
@@ -124,15 +106,7 @@ async fn shell_command_handler_to_exec_params_uses_session_shell_and_turn_contex
 
 #[test]
 fn shell_command_handler_respects_explicit_login_flag() {
-    let (_tx, shell_snapshot) = watch::channel(Some(Arc::new(ShellSnapshot {
-        path: PathBuf::from("/tmp/snapshot.sh"),
-        cwd: PathBuf::from("/tmp"),
-    })));
-    let shell = Shell {
-        shell_type: ShellType::Bash,
-        shell_path: PathBuf::from("/bin/bash"),
-        shell_snapshot,
-    };
+    let shell = Shell::new(ShellType::Bash, PathBuf::from("/bin/bash"));
 
     let login_command = ShellCommandHandler::base_command(
         &shell,
