+++ codex-rs/shell-command/src/bash.rs
@@ -5,8 +5,8 @@ use tree_sitter::Parser;
 use tree_sitter::Tree;
 use tree_sitter_bash::LANGUAGE as BASH;
 
-use crate::shell_detect::ShellType;
-use crate::shell_detect::detect_shell_type;
+use codex_shell::ShellType;
+use codex_shell::detect_shell_type;
 
 /// Parse the provided bash source using tree-sitter-bash, returning a Tree on
 /// success or None if parsing failed.
