+++ codex-rs/shell-command/src/lib.rs
@@ -1,7 +1,5 @@
 //! Command parsing and safety utilities shared across Codex crates.
 
-mod shell_detect;
-
 pub mod bash;
 pub mod command_safety;
 pub mod parse_command;
