+++ codex-rs/shell-command/src/powershell.rs
@@ -2,8 +2,8 @@ use std::path::PathBuf;
 
 use codex_utils_absolute_path::AbsolutePathBuf;
 
-use crate::shell_detect::ShellType;
-use crate::shell_detect::detect_shell_type;
+use codex_shell::ShellType;
+use codex_shell::detect_shell_type;
 
 const POWERSHELL_FLAGS: &[&str] = &["-nologo", "-noprofile", "-command", "-c"];
