+++ codex-rs/shell/BUILD.bazel
@@ -1,0 +1,7 @@
+load("//:defs.bzl", "codex_rust_crate")
+
+codex_rust_crate(
+    name = "shell",
+    crate_name = "codex_shell",
+    test_tags = ["no-sandbox"],
+)
